Improve import path detection
Use a simpler and more reliable method of finding DMD's import paths.

Instead of attempting to locate the correct dmd.conf file and parse
it correctly, invoke DMD with a non-existing filename and let it
report to us which paths it looked for the file in, which coincide
with the import paths the compiler would otherwise use.

This fixes import path detection when dmd.conf in cases such as:
- when dmd.conf is using non-trivial syntax, such as:
  - the "%@P%" macro
  - multiple DFLAGS definitions
  - multiple [Environment(...)] sections
  - external environment variables
- when dmd is invoked via a launcher script
- when the configuration file is not named dmd.conf (e.g. on
  Windows it is called sc.ini)